wip lighting

+++ src/Renderer/VkRealTimeRenderer.cpp
@@ -66,7 +66,7 @@ VkExtent2D chooseSwapExtent(const VkSurfaceCapabilitiesKHR& capabilities, int wi
 		return capabilities.currentExtent;
 	}
 	else {
-		VkExtent2D actualExtent = { width, height };
+		VkExtent2D actualExtent = { (uint32_t)width, (uint32_t)height };
 
 		actualExtent.width = std::max(capabilities.minImageExtent.width, std::min(capabilities.maxImageExtent.width, actualExtent.width));
 		actualExtent.height = std::max(capabilities.minImageExtent.height, std::min(capabilities.maxImageExtent.height, actualExtent.height));
